fix(wrapperModules.xplr): disabled -> enable
swap luaInit.<name>.disabled for enabled, and same with plugins set.

I wish `enable` to be the standard name for these options regardless of
its default value.

This is because people should not have to look up what it is called, and
contributors should not have to think about what to call it